Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,7 +2,8 @@
 
 int main() {
     int a,b,c;
+
     scanf("%d%d%d",&a,&b,&c);
-    printf("Average: %.2f\n",(a+b+c)/3);
+    printf("Average: %.2f\n",(float)(a+b+c)/3);
     return 0;
 }
